Add work-around for "duplicated directive" warning
add_directive() modifies global state in docutils.parsers.rst.directives,
even though it looks like a method on the Sphinx() object. So when
add_directive() is called multiple times, warning is emitted. This might
be an error in add_directive() design, but let's just add this hack for
now. This is only tests after all.